new route for action is added
