+++ Backend/src/routes/user.routes.js
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import {registerUser,loginUser,logoutUser, getUserProfile, followUser, suggestRelevantUsers, editProfile} from "../controllers/user.controller.js";
+import {registerUser,loginUser,logoutUser, getUserProfile, suggestRelevantUsers, editProfile} from "../controllers/user.controller.js";
 import { upload } from "../Middleware/multer.middleware.js";
 import { verifyJWT } from "../Middleware/auth.middleware.js";
 const router = Router();
@@ -7,8 +7,7 @@ const router = Router();
 router.route("/register").post(registerUser);
 router.route("/login").post(loginUser);
 router.route("/logout").post(verifyJWT, logoutUser);
-router.route("/getUserProfile").get(getUserProfile);
-router.route("/follow").post(followUser);
+router.route("/getUserProfile").get(verifyJWT,getUserProfile);
 router.route("/getUserSuggestion").get(suggestRelevantUsers);
 router.route("/editProfile").post(
     upload.fields([
